Extract a couple more utilites from run.js
