+++ packages/core/lib/commands/help/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   run: require("./run"),
-  meta: require("./meta")
+  meta: require("./meta"),
+  displayCommandHelp: require("./displayCommandHelp")
 };
